WEB DEVELOPMENT
A landing page is a standalone web page specifically designed to achieve a particular objective, often associated with marketing or advertising campaigns. Its primary goal is to convert visitors into leads or customers by encouraging them to take a specific action, such as making a purchase, signing up for a service, downloading an e-book, or subscribing to a newsletter.